console,account: Add lib to create password validation schemas
