+++ pkg/webui/console/components/user-data-form/index.js
@@ -27,7 +27,7 @@ import ModalButton from '@ttn-lw/components/button/modal-button'
 import Yup from '@ttn-lw/lib/yup'
 import PropTypes from '@ttn-lw/lib/prop-types'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
-import { hasSpecial, hasUpper, hasDigit, hasMinLength, hasMaxLength } from '@ttn-lw/lib/password'
+import createPasswordValidationSchema from '@ttn-lw/lib/create-password-validation-schema'
 
 import { id as userIdRegexp } from '@console/lib/regexp'
 
@@ -52,14 +52,9 @@ const m = defineMessages({
     'Primary email address used for logging in; this address is not publicly visible',
   modalWarning:
     'Are you sure you want to delete the user "{userId}". This action cannot be undone and it will not be possible to reuse the user ID.',
-  validateSpecial:
-    '{field} must have at least {special} special {special, plural, one {character} other {characters}}',
-  validateUppercase:
-    '{field} must have at least {upper} uppercase {upper, plural, one {character} other {characters}}',
-  validateDigit: '{field} must have at least {digit} {digit, plural, one {digit} other {digits}}',
 })
 
-const validationSchema = Yup.object().shape({
+const baseValidationSchema = Yup.object().shape({
   ids: Yup.object().shape({
     user_id: Yup.string()
       .matches(userIdRegexp, Yup.passValues(sharedMessages.validateIdFormat))
@@ -79,45 +74,6 @@ const validationSchema = Yup.object().shape({
   description: Yup.string().max(2000, Yup.passValues(sharedMessages.validateTooLong)),
 })
 
-const createPasswordValidationSchema = requirements => {
-  const passwordValidation = Yup.string()
-    .required(sharedMessages.validateRequired)
-    .test(
-      'min-length',
-      { message: sharedMessages.validateTooShort, values: { min: requirements.min_length } },
-      password => hasMinLength(password, requirements.min_length),
-    )
-    .test(
-      'max-length',
-      { message: sharedMessages.validateTooLong, values: { max: requirements.max_length } },
-      password => hasMaxLength(password, requirements.max_length),
-    )
-    .test(
-      'min-special',
-      { message: m.validateSpecial, values: { special: requirements.min_special } },
-      password => hasSpecial(password, requirements.min_special),
-    )
-    .test(
-      'min-upper',
-      { message: m.validateUppercase, values: { upper: requirements.min_uppercase } },
-      password => hasUpper(password, requirements.min_uppercase),
-    )
-    .test(
-      'min-digit',
-      { message: m.validateDigit, values: { digit: requirements.min_digits } },
-      password => hasDigit(password, requirements.min_digits),
-    )
-
-  return validationSchema.concat(
-    Yup.object().shape({
-      password: passwordValidation,
-      confirmPassword: Yup.string()
-        .required(sharedMessages.validateRequired)
-        .oneOf([Yup.ref('password'), null], sharedMessages.validatePasswordMatch),
-    }),
-  )
-}
-
 @injectIntl
 class UserForm extends React.Component {
   constructor(props) {
@@ -125,8 +81,8 @@ class UserForm extends React.Component {
 
     const { update, passwordRequirements } = props
     this.validationSchema = update
-      ? validationSchema
-      : createPasswordValidationSchema(passwordRequirements)
+      ? baseValidationSchema
+      : baseValidationSchema.concat(createPasswordValidationSchema(passwordRequirements))
     this.state = {
       error: '',
     }
@@ -177,7 +133,7 @@ class UserForm extends React.Component {
   @bind
   async handleSubmit(values, { resetForm, setSubmitting }) {
     const { onSubmit, onSubmitSuccess, onSubmitFailure } = this.props
-    const castedValues = validationSchema.cast(values)
+    const castedValues = this.validationSchema.cast(values)
     await this.setState({ error: '' })
     try {
       const result = await onSubmit(castedValues)
